Do not trigger request to get EC status if Ads is not connected.
