--- js/src/pages/settings/enhanced-conversions/disabled-enhanced-conversions-card.js
@@ -1,38 +1,0 @@
-/**
- * External dependencies
- */
-import { __ } from '@wordpress/i18n';
-import { CheckboxControl } from '@wordpress/components';
-
-/**
- * Internal dependencies
- */
-import Section from '~/components/section';
-
-/**
- * Renders the settings section for Enhanced Conversions setup.
- *
- * @fires gla_documentation_link_click with `{ context: 'setup-enhanced-conversions', link_id: 'enhanced-conversions-read-more', href: 'https://support.google.com/google-ads/answer/9888656' }`
- */
-const DisabledEnhancedConversionCard = () => {
-	return (
-		<Section.Card>
-			<Section.Card.Body>
-				<CheckboxControl
-					label={ __(
-						'Send Enhanced Conversions data to Google Ads',
-						'google-listings-and-ads'
-					) }
-					checked={ false }
-					disabled={ true }
-					help={ __(
-						'Please connect your Google Ads account in order to use Enhanced Conversions data.',
-						'google-listings-and-ads'
-					) }
-				/>
-			</Section.Card.Body>
-		</Section.Card>
-	);
-};
-
-export default DisabledEnhancedConversionCard;
