+++ js/src/pages/settings/enhanced-conversions/useEnableEnhancedConversions.js
@@ -7,6 +7,7 @@ import { useSelect } from '@wordpress/data';
  * Internal dependencies
  */
 import { STORE_KEY } from '~/data';
+import useGoogleAdsAccount from '~/hooks/useGoogleAdsAccount';
 
 const selectorName = 'getEnableEnhancedConversions';
 
@@ -18,21 +19,36 @@ const selectorName = 'getEnableEnhancedConversions';
 
 /**
  * Retrieves the enabled state and resolution status for the enhanced conversions feature.
+ * If the Google Ads connection is not established or the resolution has not finished,
+ * it returns a default state indicating that enhanced conversions are not enabled.
  *
  * @return {EnableEnhancedConversions} The data and its state.
  */
 const useEnableEnhancedConversions = () => {
-	return useSelect( ( select ) => {
-		const selector = select( STORE_KEY );
+	const { hasGoogleAdsConnection, hasFinishedResolution } =
+		useGoogleAdsAccount();
 
-		return {
-			isEnabled: selector[ selectorName ](),
-			hasFinishedResolution: selector.hasFinishedResolution(
-				selectorName,
-				[]
-			),
-		};
-	}, [] );
+	return useSelect(
+		( select ) => {
+			if ( ! hasFinishedResolution || ! hasGoogleAdsConnection ) {
+				return {
+					isEnabled: false,
+					hasFinishedResolution,
+				};
+			}
+
+			const selector = select( STORE_KEY );
+
+			return {
+				isEnabled: selector[ selectorName ](),
+				hasFinishedResolution: selector.hasFinishedResolution(
+					selectorName,
+					[]
+				),
+			};
+		},
+		[ hasGoogleAdsConnection, hasFinishedResolution ]
+	);
 };
 
 export default useEnableEnhancedConversions;
